Managing conflicts on branch. Move InstrumentType to ship_config.py
