+++ src/virtualship/expedition/__init__.py
@@ -2,7 +2,6 @@
 
 from .do_expedition import do_expedition
 from .input_data import InputData
-from .instrument_type import InstrumentType
 from .schedule import Schedule
 from .ship_config import (
     ADCPConfig,
