+++ src/virtualship/utils.py
@@ -139,8 +139,8 @@ def mfp_to_yaml(coordinates_file_path: str, yaml_output_path: str):  # noqa: D41
 
     """
     # Importing Schedule and related models from expedition module
-    from virtualship.expedition.instrument_type import InstrumentType
     from virtualship.expedition.schedule import Schedule
+    from virtualship.expedition.ship_config import InstrumentType
     from virtualship.expedition.space_time_region import (
         SpaceTimeRegion,
         SpatialRange,
